pyproject.toml
[project]
name = "docgate"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard-no-fastapi-cloud-cli]>=0.128.0",
    "peewee[postgres]>=3.19.0",
    "python-dotenv>=1.2.1",
    "sniffio>=1.3.1",
    "supertokens-python>=0.30.2",
]

[tool.ruff]
line-length = 120
indent-width = 2

[tool.ruff.lint]
# 确保 "I" (isort) 规则在启用列表中
select = ["E", "F", "I"]

[tool.ruff.lint.isort]
# 这里可以像 isort 一样自定义配置
force-single-line = false   # 是否强制单行导入
combine-as-imports = true   # 是否合并 import as

[[tool.uv.index]]
url = "https://mirrors.aliyun.com/pypi/simple/"
default = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.14.13",
    "types-peewee>=3.19.0.20260109",
]

[tool.hatch.build.targets.wheel]
packages = ["docgate"]
